[Remove Nixpkgs] move concurrency code to own file
